#507 replace slider component with tabs
